Move CAATypesSelectorDialog to CAA provider
